requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.5
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
mcp==1.1.2
openai==1.54.5
pyyaml==6.0.2
